[Chore] List 탭 목록/추가 화면 분기
