+++ app/list/RegisterScreen.js
@@ -1,0 +1,9 @@
+import { StyleSheet } from "react-native";
+
+const RegisterScreen = ({ setAddMode }) => {
+  return <></>;
+};
+
+export default RegisterScreen;
+
+const styles = StyleSheet.create({});
